+++ enpt/processors/atmospheric_correction/atmospheric_correction.py
@@ -84,7 +84,7 @@ class AtmosphericCorrector(object):
         options = self.get_ac_options(enmap_ImageL1)
         enmap_ImageL1.logger.debug('AC options: \n' + pprint.pformat(options))
 
-        enmap_ImageL1.transform_vnir_to_swir_raster('mask_water')
+        enmap_ImageL1.transform_vnir_to_swir_raster('mask_landwater')
 
         # run AC
         enmap_ImageL1.logger.info("Starting atmospheric correction for VNIR and SWIR detector. "